+++ qupath-core/src/main/java/qupath/lib/io/GsonTools.java
@@ -25,6 +25,7 @@ import java.awt.geom.AffineTransform;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
@@ -47,9 +48,12 @@ import com.google.gson.stream.JsonWriter;
 
 import qupath.lib.common.ColorTools;
 import qupath.lib.io.PathObjectTypeAdapters.FeatureCollection;
+import qupath.lib.io.PathObjectTypeAdapters.HierarchyFeatureCollection;
 import qupath.lib.measurements.MeasurementList;
 import qupath.lib.objects.PathObject;
+import qupath.lib.objects.PathRootObject;
 import qupath.lib.objects.classes.PathClass;
+import qupath.lib.objects.classes.PathClassTools;
 import qupath.lib.regions.ImagePlane;
 import qupath.lib.roi.interfaces.ROI;
 
@@ -107,11 +111,18 @@ public class GsonTools {
 		
 		@SuppressWarnings("unchecked")
 		static <T> TypeAdapter<T> getTypeAdaptor(Class<T> cls) {
+			// No point serializing the root object alone - serialize the whole hierarchy instead
+			if (PathRootObject.class.isAssignableFrom(cls))
+				return (TypeAdapter<T>)PathObjectTypeAdapters.PathObjectTypeAdapter.INSTANCE_HIERARCHY;
+			
 			if (PathObject.class.isAssignableFrom(cls))
 				return (TypeAdapter<T>)PathObjectTypeAdapters.PathObjectTypeAdapter.INSTANCE;
 
 			if (MeasurementList.class.isAssignableFrom(cls))
 				return (TypeAdapter<T>)PathObjectTypeAdapters.MeasurementListTypeAdapter.INSTANCE;
+
+			if (HierarchyFeatureCollection.class.isAssignableFrom(cls))
+				return (TypeAdapter<T>)PathObjectTypeAdapters.PathObjectCollectionTypeAdapter.INSTANCE_HIERARCHY;
 
 			if (FeatureCollection.class.isAssignableFrom(cls))
 				return (TypeAdapter<T>)PathObjectTypeAdapters.PathObjectCollectionTypeAdapter.INSTANCE;
@@ -304,11 +315,26 @@ public class GsonTools {
 	/**
 	 * Wrap a collection of PathObjects as a FeatureCollection. The purpose of this is to enable 
 	 * exporting a GeoJSON FeatureCollection that may be reused in other software.
-	 * @param pathObjects
-	 * @return
+	 * @param pathObjects a collection of path objects to store in a feature collection
+	 * @return a feature collection that can be used with {@link GsonTools}
 	 */
 	public static FeatureCollection wrapFeatureCollection(Collection<? extends PathObject> pathObjects) {
-		return new FeatureCollection(pathObjects);
+		return wrapFeatureCollection(pathObjects, false);
+	}
+	
+	/**
+	 * Wrap a collection of PathObjects as a FeatureCollection. The purpose of this is to enable 
+	 * exporting a GeoJSON FeatureCollection that may be reused in other software.
+	 * @param pathObjects a collection of path objects to store in a feature collection
+	 * @param includeChildObjects if true, include child object in the feature collection.
+	 *                            Note that no checks are made to avoid duplicate objects.
+	 * @return a feature collection that can be used with {@link GsonTools}
+	 */
+	public static FeatureCollection wrapFeatureCollection(Collection<? extends PathObject> pathObjects, boolean includeChildObjects) {
+		if (includeChildObjects)
+			return new HierarchyFeatureCollection(pathObjects);
+		else
+			return new FeatureCollection(pathObjects);			
 	}
 	
 	/**
@@ -352,11 +378,48 @@ public class GsonTools {
 				// Write in the default way
 				gson.toJson(value, PathClass.class, out);				
 			} else {
-				// Write in a simplified way, with toString() and an array of RGB values
-				var proxy = new PathClassProxy();
-				proxy.name = value.toString();
-				proxy.color = rgbToArray(value.getColor());
-				gson.toJson(proxy, PathClassProxy.class, out);	
+				out.beginObject();
+				var names = PathClassTools.splitNames(value);
+				if (names.size() == 1) {
+					out.name("name");
+					out.value(names.get(0));
+				} else {
+					out.name("names");
+					out.beginArray();
+					for (var name : names)
+						out.value(name);					
+					out.endArray();
+				}
+				var color = value.getColor();
+				if (color != null) {
+					out.name("color");
+					var alpha = ColorTools.alpha(color);
+					try {
+						if (alpha != 0 && alpha != 255)
+							out.jsonValue(String.format("[%d, %d, %d, %d]", ColorTools.red(color), ColorTools.green(color), ColorTools.blue(color), ColorTools.alpha(color)));
+						else
+							out.jsonValue(String.format("[%d, %d, %d]", ColorTools.red(color), ColorTools.green(color), ColorTools.blue(color)));
+					} catch (UnsupportedOperationException e) {
+						// TODO: Consider not trying to write json value, since it isn't always supported
+						out.beginArray();
+						out.value(ColorTools.red(color));
+						out.value(ColorTools.green(color));
+						out.value(ColorTools.blue(color));
+						if (alpha != 0 && alpha != 255)
+							ColorTools.alpha(color);
+						out.endArray();
+					}
+				}
+				out.endObject();
+				
+//				// Write in a simplified way, with toString() and an array of RGB values
+//				var proxy = new PathClassProxy();
+//				if (names.size() == 1)
+//					proxy.name = names.get(0);
+//				else
+//					proxy.names = names;
+//				proxy.color = rgbToArray(value.getColor());
+//				gson.toJson(proxy, PathClassProxy.class, out);	
 			}
 		}
 
@@ -378,18 +441,17 @@ public class GsonTools {
 				JsonObject pathClassObject = gson.fromJson(in, JsonObject.class);	
 				// Check if we have just serialized in the default way (with the usual private field name for color)
 				// This also should be used with PathClass.NULL_CLASS (which has an empty object)
+				// (It was the method used before v0.4.0)
 				if (pathClassObject.size() == 0 || pathClassObject.has("colorRGB") || pathClassObject.has("parentClass")) {
 					// Read in the default way, then replace with a singleton instance
 					PathClass pathClass = gson.fromJson(pathClassObject, PathClass.class);
 					return PathClass.getSingleton(pathClass);					
 				}
-				
 				// Check if we have a proxy object
-				if (pathClassObject.has("name")) {
+				if ((pathClassObject.has("name") || pathClassObject.has("names")) && pathClassObject.has("color")) {
 					PathClassProxy proxy = gson.fromJson(pathClassObject, PathClassProxy.class);
 					return proxy.getPathClass();
 				}
-				
 			}
 			throw new JsonParseException("Unable to parse PathClass from " + in);
 		}
@@ -397,13 +459,17 @@ public class GsonTools {
 		
 		private static class PathClassProxy {
 			
+			private List<String> names;
 			private String name;
 			private int[] color;
 			
 			private PathClass getPathClass() {
-				if (name == null)
-					return null;
-				Integer rgb = arrayToRgb(color);
+				Integer rgb = color == null ? null : arrayToRgb(color);
+				if (name == null) {
+					if (names == null || names.isEmpty())
+						return null;
+					return PathClass.fromCollection(names, rgb);
+				}
 				return PathClass.fromString(name, rgb);
 			}
 			
@@ -437,9 +503,14 @@ public class GsonTools {
 	private static Integer arrayToRgb(int[] rgb) {
 		if (rgb == null || rgb.length == 0)
 			return null;
-		if (rgb.length < 3)
+		if (rgb.length == 1)
 			return rgb[0];
-		return ColorTools.packRGB(rgb[0], rgb[1], rgb[2]);
+		if (rgb.length == 3)
+			return ColorTools.packRGB(rgb[0], rgb[1], rgb[2]);
+		if (rgb.length == 4)
+			// Alpha last in the array!
+			return ColorTools.packARGB(rgb[3], rgb[0], rgb[1], rgb[2]);
+		throw new IllegalArgumentException("RGB array should have length 0, 1, 3 or 4 - not " + rgb.length);
 	}