+++ qupath-core/src/main/java/qupath/lib/objects/PathObject.java
@@ -203,7 +203,7 @@ public abstract class PathObject implements Externalizable {
 		if (measurementsMap == null) {
 			synchronized(this) {
 				if (measurementsMap == null)
-					measurementsMap = measurements.asMap();
+					measurementsMap = getMeasurementList().asMap();
 			}
 		}
 		return measurementsMap;